+++ app/drivers/sensor/pinnacle/pinnacle.h
